// src/generated/contracts/domain/PrinterStatusBackfill.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsSafeInteger } from "../event/JsSafeInteger";
import type { PrinterStatusRow } from "./PrinterStatusRow";
import type { StatusCacheWarning } from "./StatusCacheWarning";

export type PrinterStatusBackfill = { streamId: string, snapshotSequence: JsSafeInteger, statuses: Array<PrinterStatusRow>, cacheWarnings: Array<StatusCacheWarning>, };